Changed LSH to an Abstract approach
