+++ LDLSH/src/main/java/SystemLayer/Data/LSHHashImpl/JavaMinHashNgrams.java
@@ -8,19 +8,26 @@ import java.util.Set;
 
 public class JavaMinHashNgrams extends JavaMinHash{
 
-    private static final String vector_dimensions = "VECTOR_DIMENSIONS";
-
-    public JavaMinHashNgrams(DataObject dataObject, int n_blocks, DataContainer dataContainer) throws Exception {
-        super(dataObject, n_blocks, dataContainer);
+    public static void setupMinHash(
+            double error,
+            int vector_dimensions,
+            long seed,
+            int level
+    ) { //l-n+1
+        int ngramDimensions = vector_dimensions - level + 1;
+        JavaMinHash.setupMinHash(error,ngramDimensions,seed);
     }
 
-    public JavaMinHashNgrams(DataContainer dataContainer) {
-        super(dataContainer);
+    public JavaMinHashNgrams( DataContainer appContext ){
+        super(appContext);
     }
 
     @Override
     public void setObject( DataObject object, int n_blocks ){
-
+        byte[] data = object.toByteArray();
+        //Calculate ngram vector
+        this.data = getSignature(data);
+        this.blocks = createBlocks(this.data, n_blocks);
     }
 
 }
