+++ LDLSH/src/main/java/SystemLayer/Data/LSHHashImpl/LSHHash.java
@@ -2,34 +2,33 @@ package SystemLayer.Data.LSHHashImpl;
 
 import SystemLayer.Data.DataObjectsImpl.DataObject;
 import java.io.Serializable;
-import java.util.Arrays;
 
 public interface LSHHash extends Serializable, Comparable<LSHHash> {
+
+    /**
+     * Sets a given object to associate with a resulting signature
+     * @param object object to be associated
+     * @param n_blocks number of divisions on the signature
+     */
     void setObject( DataObject object, int n_blocks );
+
+    /**
+     * Returns the generated complete signature
+     * @return a bite array containing the associated object's signature
+     */
     byte[] getSignature();
 
-    LSHHashBlock[] getBlocks();
-    LSHHashBlock getBlockAt( int position );
+    /**
+     * Returns the list of signature blocks
+     * @return Array of signature blocks
+     */
+    LSHHashImpl.LSHHashBlock[] getBlocks();
 
-    record LSHHashBlock( byte[] lshBlock ) implements Comparable<LSHHashBlock>{
-        @Override
-        public int compareTo(LSHHashBlock o) {
-            return Arrays.compare(lshBlock, o.lshBlock);
-        }
+    /**
+     * Returns a block associated with a given position
+     * @param position pretended block's position
+     * @return pretended block
+     */
+    LSHHashImpl.LSHHashBlock getBlockAt(int position );
 
-        @Override
-        public boolean equals(Object obj) {
-            if (obj.getClass() != LSHHashBlock.class)
-                return false;
-            LSHHashBlock conv = (LSHHashBlock) obj;
-            return Arrays.equals( this.lshBlock, conv.lshBlock );
-        }
-
-        @Override
-        public int hashCode() {
-            return Arrays.hashCode(this.lshBlock);
-        }
-
-        /**/
-    }
 }
